dt2w - handle slot change more effectively
The touch screen informs EV_SYN before it changes the slot with ABS_MT_SLOT,
however we do not consider/handle EV_SYN but we simply reset the touch if we see
the change of slot.
If a touch has already recognized and resetting it while changing the slot,
makes the touch unrecognized and required two more touches for dt2w.
Let us atleast handle a case where we do not reset the touch if we have already
recognized one touch.